+++ .travis.yml
@@ -1,12 +1,21 @@
 language: php
 php:
-  - '5.4'
-  - '5.5'
   - '5.6'
   - '7.0'
   - '7.1'
   - '7.2'
+  - '7.3'
   - 'nightly'
+
+matrix:
+  matrix:
+  include:
+    - php: '5.4'
+      dist: 'trusty'
+    - php: '5.5'
+      dist: 'trusty'
+  allow_failures:
+    - php: 'nightly'
 
 script: vendor/bin/phpunit --configuration ./build/travis-ci.xml
